Bugfixes: 1) assert(len(sequence) < 256); 2) GC and primer not being used.
Changed the model to not add PlasmidPrimers atomically. Unfortunately, this was the easiest fix for a weird problem - getting the DB value back from a trigger which sets a PK value. For non-PK values, this should work fine but SQLAlchemy treats the PK case differently. There are better solutions but this was a quick one.